salient/Python/python-chatbot-starter-v1/chatbot_starter/requirements.txt
autopep8==1.5
Django==4.2.2
django-browser-reload==1.9.0
django-multiupload==0.6.1
django-tailwind==3.5.0
langchain==0.0.187
openai==0.27.7
pinecone-client==2.2.1
pycodestyle==2.6.0
pypdf==3.9.0
python-dotenv==1.0.0
virtualenv==20.0.17
